+++ History.md
@@ -2,6 +2,15 @@ Unreleased
 ==================
 * added; `SetupFeeAccountingCode` to `Plan`
 
+* added; `AppliesToNonPlanCharges` attribute `Coupon`
+* added; `Coupons` to `Subscription`
+* added; `uuid` to `CouponRedemption`
+* added; `GetActiveRedemptions` to `Account`
+* added; `X-Api-Version` to `2.1`
+* added; `RedemptionResource` to `Coupon`
+* added; `MaxRedemptionsPerAccount` to `Coupon`
+
+
 1.2.3 (stable) / 2015-08-14
 ==================